requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
plotly
textblob